Move livestream and slido to separate component
